backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg[binary]
python-dotenv
channels
channels_redis
channels[daphne]